adding helper functions for form validation
