+++ src/components/signupform/SignupForm.js
@@ -2,6 +2,7 @@ import { useState } from "react";
 import CustomizableButton from "../common/CustomizableButton";
 import styles from "./signupform.module.css";
 import { useAuthSignUp } from "@/composables/authSignUp";
+import regexValidation from "./helperFunctions/regexValidation";
 
 function Form() {
   const [user, setUser] = useState({
@@ -20,56 +21,21 @@ function Form() {
   const handleSubmit = (e) => {
     e.preventDefault();
 
-    if (validateForm()) {
+    const errors = regexValidation(user);
+
+    if (Object.keys(errors).length === 0) {
       signup(user);
 
       setTimeout(() => {
         resetForm();
       }, 1000);
+    } else {
+      setError(errors);
     }
   };
 
   const resetForm = () => {
     setUser({ name: "", email: "", password: "" });
-  };
-
-  // Validate Submit Form
-  const validateForm = () => {
-    const errors = {};
-
-    // Validate Name
-    const namePattern = /^[a-zA-Z]{2,50}(?:\s+[a-zA-Z]{2,50})+$/;
-    if (!user.name) {
-      errors.name = "Please enter your name";
-    } else if (user.name.length < 2 || user.name.length > 50) {
-      errors.name = "Name must be between 2 and 50 characters";
-    } else if (!user.name.match(namePattern)) {
-      errors.name = "Please enter a valid name";
-    }
-
-    // Valid Email
-    const emailPattern = /^\w+@[a-zA-Z_]+?\.[a-zA-Z]{2,3}$/;
-    if (!user.email) {
-      errors.email = "Please enter your email address";
-    } else if (!user.email.match(emailPattern)) {
-      errors.email = "Please enter a valid email address";
-    }
-
-    // Valid Password
-    const passwordPattern =
-      /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[^a-zA-Z0-9])(?!.*\s).{8,20}$/;
-    if (!user.password) {
-      errors.password = "Please enter a password";
-    } else if (user.password.length < 8 || user.password.length > 20) {
-      errors.password = "Please enter a password between 8 and 20 characters";
-    } else if (!user.password.match(passwordPattern)) {
-      errors.password =
-        "Password must contain one digit, one lowercase letter, one uppercase letter, and one special character";
-    }
-
-    setError(errors);
-
-    return Object.keys(errors).length === 0;
   };
 
   return (
